modularizing web3 + explaining auth init
