+++ docker/docker-compose.yml
@@ -10,7 +10,7 @@ node:
     - --deterministic
     - --db=/ganache_data
     - --mnemonic
-    - 'minimum symptom minute gloom tragic situate silver mechanic salad amused elite beef'
+    - 'live genius aerobic observe remain husband allow main disorder protect oak powder'
     - --networkId
     - '5777'
     - --hostname
