+++ monedero-react/src/services/billeteraService.js
@@ -1,5 +1,6 @@
 import { usuarioService } from "./usuarioService"
-import { handleRPCError, walletContract } from "./web3Wrapper"
+import { walletContract } from "./blockchainService"
+import { handleRPCError } from "./errorHandling"
 
 class BilleteraService {
   async poner(monto) {
