--- monedero-react/src/services/cuentas.js
@@ -1,18 +1,0 @@
-// Hay que sacarlas de Ganache
-export const cuentas = [
-  {
-      address: '0xE9BbA9735f430156eB563C793Dc53b8F42C783DE',
-      balance: 0,
-      username: 'dodain'
-  },
-  {
-      address: '0x59283dd5EBF26705f135A3d0d7dDc9deEA45ef68',
-      balance: 0,
-      username: 'juan'
-  },
-  {
-      address: '0x3A5123002c4546dE724C79B10F610930B2Ec2207',
-      balance: 0,
-      username: 'dini'
-  },
-]
